fix(shields): stabilize Hermes MCP state mutation
