+++ test/helpers/hermes-shields-provider-consumer-harness.ts
@@ -341,6 +341,11 @@ export function createHermesShieldsProviderConsumerHarness(
       ) {
         return `${SEALED_PLAN_HELP}\n`;
       }
+      const gatewayControlIndex = command.indexOf("/usr/local/bin/nemoclaw-gateway-control");
+      if (gatewayControlIndex >= 0 && command[gatewayControlIndex + 1] === "restart") {
+        const nonce = String(command[gatewayControlIndex + 2]);
+        return `v1 ${nonce} complete ok 10 11\nGATEWAY_PID=11\n`;
+      }
       if (command[0] === "stat" && command.at(-1) === hermesProviderConsumerTarget.configDir) {
         return "3770 sandbox:sandbox\n";
       }
